add ignore_http_codes put param
takes an array of http response codes that the put will not consider an
error. In one case this is useful because Artifactory currently
returns "404" if the uploaded file already exists but won't be overridden.

This will therefore allow us to count this as a success if the upload
was previously run successfully.